Create Convert Email to SNMP Trap
Piece of code taken from Office Get Outlook Inbox email to have the script read the mail.

 

The script reads the inbox items and then processes the required mails (to be converted to trap) based on the given criteria (here, I am matching the subject).

 

Pre-Requisites to get this script work:

PowerShell and Outlook Client need to be installed on a windows server. Outlook needs to be configured with the mailbox of the mail ID to which you get the mail (Example: add the mailbox for abcd@xyz.com if you get the mails to this mail ID)

Also, set Outlook to allow the script to read the inbox. For this, open Outlook --> File --> Options --> Trust Center --> Trust Center Settings --> Programmatic Access --> Select the radio button for "Never warn me about suspicious activity (not recommended).

Otherwise, the script will wait for the outlook program to allow it.